feat : 설치된 오브제 좌표, 회전, 크기 값 추가

See merge request s09-final/S09P31B304!132

+++ thingdong/src/main/java/com/bell/thingdong/domain/objet/dto/ObjectSizeDto.java
@@ -12,8 +12,8 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class ObjectSizeDto {
 	@Schema(description = "오브젝트 높이 값", example = "2")
-	private Double height;
+	private Long height;
 
 	@Schema(description = "오브젝트 넓이 값", example = "3")
-	private Double width;
+	private Long width;
 }
